% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CalcCorVector.R
\name{CalcCorVector}
\alias{CalcCorVector}
\title{Calculate correlation vector}
\usage{
CalcCorVector(gene.vec, gem)
}
\arguments{
\item{gene.vec}{}

\item{gem}{}
}
\value{
Correlation vector
}
\description{
Calculate correlation vector
}

