% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CCLE_copy}
\alias{CCLE_copy}
\title{Log ratio copynumber of CCLE data}
\format{A data frame with 23224 rows and 976 variables:
\itemize{
  \item geneName: HGNC approved gene name
  \item NumChr: Chromosome number
  \item txStart: Start position on chromosome
  \item txEnd: End position on chromosome
  \item X1321N1_CENTRAL_NERVOUS_SYSTEM: Log ratio copynumber on sample X1321N1_CENTRAL_NERVOUS_SYSTEM
  \item X143B_BONE: Log ratio copynumber on sample X143B_BONE
  \item ...
}}
\usage{
CCLE_copy

CCLE_copy
}
\description{
A dataset containing the log ratio copynumber for over 20000 genes and nearly 1000 samples.
Availiable from the broad institute: http://www.broadinstitute.org/ccle/data/browseData
Filename: CCLE_copynumber_byGene_2012-04-06.txt

A dataset containing the gene-centric RMA-normalized mRNA expression data for nearly 20000 genes and nearly 1000 samples.
Availiable from the broad institute: http://www.broadinstitute.org/ccle/data/browseData
Filename: CCLE_Expression_Entrez_2012-04-06.gct.gz
}
\keyword{datasets}

