% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CCLE_samples}
\alias{CCLE_samples}
\title{Clinical information for CCLE data}
\format{A data frame with 967 rows and 14 variables:
\itemize{
  \item CCLE.name: Sample name identifier.
  \item Cell.line.primary.name: Cell line name.
  \item Cell.line.aliases: Any known aliases of cell line.
  \item Gender: Gender of patient cell line derived from.
  \item Site.Primary: Primary site cell line derived from.
  \item Histology: Histology of tumour cell line derived from.
  \item Hist.Subtype1: Histology subtype of tumour cell line derived from.
  \item Notes: Additional notes.
  \item Source: Source of the cell line.
  \item Expression.arrays: Expression array used.
  \item SNP.arrays: SNP array used.
  \item Oncomap: Oncomap mutation array used.
  \item Hybrid.Capture.Sequencing: Hybrid capture sequencing used.
  \item Name: Sample name identifier
}}
\source{
\url{http://www.broadinstitute.org/ccle/data/browseData} Filename: CCLE_sample_info_file_2012-04-06.txt
}
\usage{
CCLE_samples
}
\value{
NA
}
\description{
A dataset containing clinical information for the CCLE samples.
}
\keyword{datasets}
