% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CorScoreCalc.r
\name{CorScoreCalc}
\alias{CorScoreCalc}
\title{Calculate correlation score}
\usage{
CorScoreCalc(gene.expr.matrix, sample.vec)
}
\arguments{
\item{gene.expr.matrix}{}

\item{sample.vec}{}
}
\value{
The correlation score
}
\description{
Calculate correlation score
}

