% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVEval.R
\name{CVEval}
\alias{CVEval}
\title{Calculate correlation vector}
\usage{
CVEval(gem.part, gem.all, seed, splits)
}
\arguments{
\item{gem.part}{}
}
\value{
Correlation vector
}
\description{
Calculate correlation vector
}

