% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindSeed.R
\name{FindSeed}
\alias{FindSeed}
\title{Find highly correlated seed of samples for gene expression matrix}
\usage{
FindSeed(gem, seed.size, iterations, initial.seed = NULL, messages = 100)
}
\arguments{
\item{gem}{Gene expression matrix}

\item{seed.size}{Size of sample seed}

\item{iterations}{Number of iterations}

\item{initial.seed}{Initial seed used, if NULL randomly chosen}

\item{messages}{frequency of progress messages}
}
\value{
Highly correlated seed
}
\description{
Find highly correlated seed of samples for gene expression matrix
}
\examples{
data(CCLE_small)
data(Mitochondrial_genes)

mito.loc <- which(row.names(CCLE_small) \%in\% Mitochondrial_genes)
CCLE.mito <- CCLE_small[mito.loc,]

set.seed(101)
CCLE.seed <- FindSeed(gem = CCLE.mito,
                      seed.size = 10,
                      iterations = 100,
                      messages = 100)

CCLE.cor.vec <- CVEval(gem.part = CCLE.mito,
                       gem.all = CCLE_small,
                       seed = CCLE.seed, splits = 10)

GEA <- GOEnrichmentAnalysis(gene.names = row.names(CCLE_small),
                            gene.values = CCLE.cor.vec,
                            sig.rate = 0.05)

}

