% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindSeedGroups.R
\name{FindSeedGroups}
\alias{FindSeedGroups}
\title{FindSeed version for when pattern of regulation is known}
\usage{
FindSeedGroups(gem, seed.size, iterations, group1.loc, group2.loc,
  initial.seed = NULL, full.detail = FALSE)
}
\arguments{
\item{gem}{}
}
\value{
Seed of highly correlating samples
}
\description{
FindSeed version for when pattern of regulation is known
}

