% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ForkClassifier.R
\name{ForkClassifier}
\alias{ForkClassifier}
\title{Classification of fork status}
\usage{
ForkClassifier(pc1, samp.num)
}
\arguments{
\item{pc1}{}

\item{samp.num}{}
}
\value{
Classification of each sample
}
\description{
Classification of fork status
}

