% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneVecFunCalc.R
\name{GeneVecFunCalc}
\alias{GeneVecFunCalc}
\title{Function needed for gene vector calculation}
\usage{
GeneVecFunCalc(gem, seed, n)
}
\arguments{
\item{gem}{Gene set expression matrix}

\item{seed}{Seed of highly correlating samples}

\item{n}{Number of splits}
}
\value{
Calculation
}
\description{
Function needed for gene vector calculation
}

