% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GO_term_EG}
\alias{GO_term_EG}
\title{List of entrez gene names in GO terms}
\format{A list of GO terms with corresponding entrez gene names:}
\usage{
GO_term_EG
}
\description{
List of entrez gene names in GO terms
}
\keyword{datasets}

