% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GO_term_genes}
\alias{GO_term_genes}
\title{List of HGNC gene names in GO terms}
\format{A list of GO terms with corresponding gene names:}
\usage{
GO_term_genes
}
\description{
List of HGNC gene names in GO terms
}
\keyword{datasets}

