% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GO_term_matrix}
\alias{GO_term_matrix}
\title{GO terms ID and description}
\format{A data frame with 18826 rows and 2 variables:
 \itemize{
  \item GOID: GO ID number
  \item TERM: GO Term description
}}
\usage{
GO_term_matrix
}
\description{
A dataset containing over 18000 GO terms with their ID number
}
\keyword{datasets}

