% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoDataLoad.R
\name{GoDataLoad}
\alias{GoDataLoad}
\title{Load data for Gene set enrichment analysis}
\usage{
GoDataLoad()
}
\value{
List of GO terms and genes in each GO term
}
\description{
Load data for Gene set enrichment analysis
}

