% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOEnrichmentAnalysis.R
\name{GOEnrichmentAnalysis}
\alias{GOEnrichmentAnalysis}
\title{Calculate gene set enrichment of correlation vector using Mann-Whitney test}
\usage{
GOEnrichmentAnalysis(gene.names, gene.values, sig.rate)
}
\arguments{
\item{gene.names}{Names of the genes}

\item{gene.values}{Values associated with the genes}

\item{sig.rate}{Level of significance}

\item{output.type}{Alternative output}
}
\value{
Significant gene sets
}
\description{
Calculate gene set enrichment of correlation vector using Mann-Whitney test
}

