% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCbiclust.R
\docType{package}
\name{MCbiclust}
\alias{MCbiclust}
\alias{MCbiclust-package}
\title{MCbiclust: Massively Correlated biclustering}
\description{
MCbiclust is a R package for running massively correlating biclustering
analysis.MCbiclust aims to find large scale biclusters with selected 
features being highly correlated with each other over a subset of samples.
}
\details{
The package was originally designed in order to solve a problem in
bioinformatics: to find biclusters representing different modes of regulation
of mitochondria gene expression in disease states such as breast cancer. 
The same methods however, can be used on any gene expression data set to
find biclusters of interest.

To learn more about MCbiclust, start with the vignette:
\code{browseVignettes(package = "MCbiclust")}
}

