% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Mitochondrial_genes}
\alias{Mitochondrial_genes}
\title{List of known mitochondrial genes}
\format{A Character vector of the HGNC approved gene names:}
\usage{
Mitochondrial_genes
}
\description{
A dataset from MitoCarta1.0 containing the 1023 mitochondrial genes
Availiable from the broad institute: http://www.broadinstitute.org/scientific-community/science/programs/metabolic-disease-program/publications/mitocarta/mitocarta-in-0
}
\keyword{datasets}

