% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PC1Align.R
\name{PC1Align}
\alias{PC1Align}
\title{Make sure PC1 and Correlation Vector are aligned}
\usage{
PC1Align(gem, PC1, CV, bic)
}
\arguments{
\item{gem}{Gene expression matrix}

\item{PC1}{First principal component}

\item{CV}{Correlation vector}

\item{bic}{bicluster}
}
\value{
Aligned PC1 vector
}
\description{
Make sure PC1 and Correlation Vector are aligned
}

