% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PC1VecFun.R
\name{PC1VecFun}
\alias{PC1VecFun}
\title{Calculate PC1 of found pattern}
\usage{
PC1VecFun(top.gem, seed.sort, n)
}
\arguments{
\item{top.gem}{}

\item{seed.sort}{}

\item{n}{}
}
\value{
PC1 value for each sample
}
\description{
Calculate PC1 of found pattern
}

