% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointScoreCalc.R
\name{PointScoreCalc}
\alias{PointScoreCalc}
\title{Calculate point score for samples based on gene groups}
\usage{
PointScoreCalc(gem, gloc1, gloc2)
}
\arguments{
\item{gem}{}
}
\value{
Numeric vector of point score for each sample
}
\description{
Calculate point score for samples based on gene groups
}

