% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleSortGroups.R
\name{SampleSortGroups}
\alias{SampleSortGroups}
\title{Sort all samples by strength to a known correlation pattern}
\usage{
SampleSortGroups(gem, group1.loc, group2.loc, seed, num.cores = NULL,
  sort.length = NULL)
}
\arguments{
\item{gem}{Gene expression matrix}

\item{group1.loc}{Location of genes highly correlated to each other but anti-correlated to group2}

\item{group2.loc}{Location of genes highly correlated to each other but anti-correlated to group1}

\item{seed}{Seed for highly correlating samples}

\item{num.cores}{Number of cores used in parallel evaluation}

\item{sort.length}{Number of samples to be sorted}
}
\value{
Order of samples by strength to correlation pattern
}
\description{
Sort all samples by strength to a known correlation pattern
}

