% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SeedGenerator.R
\name{SeedGenerator}
\alias{SeedGenerator}
\title{Calculate initial seeds to minimise overlaps}
\usage{
SeedGenerator(seed.size, numbers, sample.length, break.num = 0,
  attempts = 1000)
}
\arguments{
\item{seed.size}{}

\item{numbers}{}

\item{sample.length}{}

\item{break.num}{}

\item{attempts}{}
}
\value{
Initial seed list
}
\description{
Calculate initial seeds to minimise overlaps
}

