% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SilhouetteClustGroups.R
\name{SilhouetteClustGroups}
\alias{SilhouetteClustGroups}
\title{Slihouette validation of correlation vector clusters}
\usage{
SilhouetteClustGroups(cor.vec.mat, max.clusters, plots = FALSE)
}
\arguments{
\item{cor.vec.mat}{}

\item{max.clusters}{}

\item{plots}{}
}
\value{
The distinct clusters of correlation vectors
}
\description{
Slihouette validation of correlation vector clusters
}

