% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Vignette_sort}
\alias{Vignette_sort}
\title{List of samples pre-sorted for vignette}
\format{A list of the different sample sortings calculated in the vignette:}
\usage{
Vignette_sort
}
\value{
NA
}
\description{
Different sortings of CCLE_small calcualted in the Vignette.
}
\keyword{datasets}

