% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CalcCorScoreVec.R
\name{CalcCorScoreVec}
\alias{CalcCorScoreVec}
\title{Calculate numeric vector showing correlation score decrease with sample order}
\usage{
CalcCorScoreVec(ordered.list, gem, seed.size = 10, mc = FALSE,
  num.cores = NULL)
}
\arguments{
\item{ordered.list}{}

\item{gem}{}

\item{seed.size}{}

\item{mc}{}

\item{num.cores}{}
}
\value{
Vector of correlation scores
}
\description{
Calculate numeric vector showing correlation score decrease with sample order
}

