% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcCorVector.R
\name{CalcCorVector}
\alias{CalcCorVector}
\title{Calculate correlation vector}
\usage{
CalcCorVector(gene.vec, gem)
}
\arguments{
\item{gene.vec}{Single vector containing average expression levels}

\item{gem}{Gene expression matrix}
}
\value{
Correlation vector
}
\description{
Calculate correlation vector
}

