% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorScoreCalc.r
\name{CorScoreCalc}
\alias{CorScoreCalc}
\title{Calculate correlation score}
\usage{
CorScoreCalc(gene.expr.matrix, sample.vec)
}
\arguments{
\item{gene.expr.matrix}{Gene expression matrix}

\item{sample.vec}{Vector of samples}
}
\value{
The correlation score
}
\description{
Calculate correlation score
}

