% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVEval.R
\name{CVEval}
\alias{CVEval}
\title{Calculate correlation vector}
\usage{
CVEval(gem.part, gem.all, seed, splits)
}
\arguments{
\item{gem.part}{Part of gene expression matrix only containing gene set of interest}

\item{gem.all}{All of gene expression matrix}

\item{seed}{Seed of highly correlating samples}

\item{splits}{Number of cuts from hierarchical clustering}
}
\value{
Correlation vector
}
\description{
Calculate correlation vector
}
\examples{
data(CCLE_small)
data(Mitochondrial_genes)

mito.loc <- which(row.names(CCLE_small) \%in\% Mitochondrial_genes)
CCLE.mito <- CCLE_small[mito.loc,]

set.seed(101)
CCLE.seed <- FindSeed(gem = CCLE.mito,
                      seed.size = 10,
                      iterations = 100,
                      messages = 100)

CCLE.cor.vec <- CVEval(gem.part = CCLE.mito,
                       gem.all = CCLE_small,
                       seed = CCLE.seed, splits = 10)

GEA <- GOEnrichmentAnalysis(gene.names = row.names(CCLE_small),
                            gene.values = CCLE.cor.vec,
                            sig.rate = 0.05)

}

