% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVEval.R
\name{CVEval}
\alias{CVEval}
\title{Calculate correlation vector}
\usage{
CVEval(gem.part, gem.all, seed, splits)
}
\arguments{
\item{gem.part}{Part of gene expression matrix only containing gene set of interest}

\item{gem.all}{All of gene expression matrix}

\item{seed}{Seed of highly correlating samples}

\item{splits}{Number of cuts from hierarchical clustering}
}
\value{
Correlation vector
}
\description{
Calculate correlation vector
}

