% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVPlot.R
\name{CVPlot}
\alias{CVPlot}
\title{Make correlation vector plot}
\usage{
CVPlot(cv.df, geneset.loc, geneset.name, alpha1 = 0.005, alpha2 = 0.1,
  cnames = NULL)
}
\arguments{
\item{cv.df}{A dataframe containing the correlation vectors of one or more patterns.}

\item{geneset.loc}{A gene set of interest (e.g. mitochondrial) to be plotted separately from rest of genes.}

\item{alpha1}{Transparency level of non-gene set genes}

\item{alpha2}{Transparency level of gene set genes}

\item{cnames}{Character vector containing names for the correlation vector}
}
\value{
A plot of the correlation vectors
}
\description{
Make correlation vector plot
}

