% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindSeed.R
\name{FindSeed}
\alias{FindSeed}
\title{Find highly correlated seed of samples for gene expression matrix}
\usage{
FindSeed(gem, seed.size, iterations, initial.seed = NULL)
}
\arguments{
\item{gem}{Gene expression matrix}

\item{seed.size}{Size of sample seed}

\item{iterations}{Number of iterations}

\item{initial.seed}{Initial seed used, if NULL randomly chosen}
}
\value{
Highly correlated seed
}
\description{
Find highly correlated seed of samples for gene expression matrix
}

