% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FindSeed.R
\name{FindSeed}
\alias{FindSeed}
\title{Find highly correlated seed of samples for gene expression matrix}
\usage{
FindSeed(gem, seed.size, iterations, initial.seed = NULL,
  full.detail = FALSE)
}
\arguments{
\item{gem}{}

\item{seed.size}{}

\item{iterations}{}

\item{initial.seed}{}

\item{full.detail}{}
}
\value{
Highly correlated seed
}
\description{
Find highly correlated seed of samples for gene expression matrix
}

