% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindSeedGroups.R
\name{FindSeedGroups}
\alias{FindSeedGroups}
\title{FindSeed version for when pattern of regulation is known}
\usage{
FindSeedGroups(gem, seed.size, iterations, group1.loc, group2.loc,
  initial.seed = NULL, full.detail = FALSE)
}
\arguments{
\item{gem}{Gene expression matrix}

\item{seed.size}{Number of samples in seed}

\item{iterations}{Number of iterations}

\item{group1.loc}{Location of genes highly correlated to each other but anti-correlated to group2}

\item{group2.loc}{Location of genes highly correlated to each other but anti-correlated to group1}

\item{initial.seed}{Which seed to begin the algorithm on, randomly selected if NULL}

\item{full.detail}{Return additional information}
}
\value{
Seed of highly correlating samples
}
\description{
FindSeed version for when pattern of regulation is known
}

