% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FindSeedGroups.R
\name{FindSeedGroups}
\alias{FindSeedGroups}
\title{FindSeed version for when pattern of regulation is known}
\usage{
FindSeedGroups(gem, seed.size, iterations, group1.loc, group2.loc,
  initial.seed = NULL, full.detail = FALSE)
}
\arguments{
\item{gem}{}

\item{seed.size}{}

\item{iterations}{}

\item{group1.loc}{}

\item{group2.loc}{}

\item{initial.seed}{}

\item{full.detail}{}
}
\value{
Seed of highly correlating samples
}
\description{
FindSeed version for when pattern of regulation is known
}

