% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForkClassifier.R
\name{ForkClassifier}
\alias{ForkClassifier}
\title{Classification of fork status}
\usage{
ForkClassifier(pc1, samp.num)
}
\arguments{
\item{pc1}{PC1 values of samples}

\item{samp.num}{Number of samples in the bicluster}
}
\value{
Classification of each sample
}
\description{
Classification of fork status
}

