% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FSGZeroRowUpdate.R
\name{FSGZeroRowUpdate}
\alias{FSGZeroRowUpdate}
\title{FindSeedGroups function for updating rows with 0 sd}
\usage{
FSGZeroRowUpdate(gem, seed, group1.loc, group2.loc)
}
\arguments{
\item{gem}{}

\item{seed}{}

\item{group1.loc}{}

\item{group2.loc}{}
}
\value{
Updated gene locations and correlation matrix
}
\description{
FindSeedGroups function for updating rows with 0 sd
}

