% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GeneVecFun.R
\name{GeneVecFun}
\alias{GeneVecFun}
\title{Calcuclates gene vector used in calculation of correlation vector}
\usage{
GeneVecFun(gem, seed, splits)
}
\arguments{
\item{gem}{}

\item{seed}{}

\item{splits}{}
}
\value{
Average expression vector that matches pattern in seed
}
\description{
Calcuclates gene vector used in calculation of correlation vector
}

