% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneVecFun.R
\name{GeneVecFun}
\alias{GeneVecFun}
\title{Calcuclates gene vector used in calculation of correlation vector}
\usage{
GeneVecFun(gem, seed, splits)
}
\arguments{
\item{gem}{Gene expression matrix}

\item{seed}{Seed of highly correlating samples}

\item{splits}{Maximum number of cuts from hierarchical clustering}
}
\value{
Average expression vector that matches pattern in seed
}
\description{
Calcuclates gene vector used in calculation of correlation vector
}

