% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneVecFunCalc.R
\name{GeneVecFunCalc}
\alias{GeneVecFunCalc}
\title{Function needed for gene vector calculation}
\usage{
GeneVecFunCalc(gem, seed, n)
}
\value{
Calculation
}
\description{
Function needed for gene vector calculation
}

