% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOEnrichmentAnalysis.R
\name{GOEnrichmentAnalysis}
\alias{GOEnrichmentAnalysis}
\title{Calculate gene set enrichment of correlation vector using Mann-Whitney test}
\usage{
GOEnrichmentAnalysis(gene.names, gene.values, sig.rate)
}
\arguments{
\item{gene.names}{Names of the genes}

\item{gene.values}{Values associated with the genes}

\item{sig.rate}{Level of significance}
}
\value{
Significant gene sets
}
\description{
Calculate gene set enrichment of correlation vector using Mann-Whitney test
}
\examples{
data(CCLE_data)
data(Mitochondrial_genes)

CCLE.data <- CCLE_data[,-c(1,2)]
mito.loc <- which(as.character(CCLE_data[,2]) \%in\% Mitochondrial_genes)
CCLE.mito <- CCLE_data[mito.loc,-c(1,2)]
row.names(CCLE.mito) <- CCLE_data[mito.loc,2]

set.seed(101)
CCLE.seed <- FindSeed(gem = CCLE.mito,
                      seed.size = 10,
                      iterations = 100,
                      messages = 100)

CCLE.cor.vec <- CVEval(gem.part = CCLE.mito,
                       gem.all = CCLE.data,
                       seed = CCLE.seed, splits = 10)

GEA <- GOEnrichmentAnalysis(gene.names = CCLE_data[,2],
                            gene.values = CCLE.cor.vec,
                            sig.rate = 0.05)

}

