% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GOEnrichmentAnalysis.R
\name{GOEnrichmentAnalysis}
\alias{GOEnrichmentAnalysis}
\title{Calculate gene set enrichment of correlation vector using Mann-Whitney test}
\usage{
GOEnrichmentAnalysis(gene.names, gene.values, sig.rate)
}
\arguments{
\item{gene.names}{}

\item{gene.values}{}

\item{sig.rate}{}
}
\value{
Significant gene sets
}
\description{
Calculate gene set enrichment of correlation vector using Mann-Whitney test
}

