% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOEnrichmentAnalysis.R
\name{GOEnrichmentAnalysis}
\alias{GOEnrichmentAnalysis}
\title{Calculate gene set enrichment of correlation vector using Mann-Whitney test}
\usage{
GOEnrichmentAnalysis(gene.names, gene.values, sig.rate, output.type = NULL)
}
\value{
Significant gene sets
}
\description{
Calculate gene set enrichment of correlation vector using Mann-Whitney test
}

