% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HclustGenesHiCor.R
\name{HclustGenesHiCor}
\alias{HclustGenesHiCor}
\title{Find the most highly correlated genes using hierarchical clustering}
\usage{
HclustGenesHiCor(gem, seed, cuts)
}
\arguments{
\item{gem}{Gene expression matrix}

\item{seed}{Seed of highly correlating samples}

\item{cuts}{Number of groups to cut dendogram into}
}
\value{
Numeric vector of most highly correlated genes
}
\description{
Find the most highly correlated genes using hierarchical clustering
}
\examples{
data(CCLE_data)
data(Mitochondrial_genes)

mito.loc <- which(row.names(CCLE_small) \%in\% Mitochondrial_genes)
CCLE.mito <- CCLE_small[mito.loc,]

random.seed <- sample(seq(length = dim(CCLE.mito)[2]),10)
CCLE.seed <- FindSeed(gem = CCLE.mito,
                      seed.size = 10,
                      iterations = 100,
                      messages = 100)


CorScoreCalc(CCLE.mito, random.seed)
CorScoreCalc(CCLE.mito, CCLE.seed)

CCLE.hicor.genes <- as.numeric(HclustGenesHiCor(CCLE.mito,
                                                CCLE.seed,
                                                cuts = 8))

CorScoreCalc(CCLE.mito[CCLE.hicor.genes,], CCLE.seed)
}

