% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HclustGenesHiCor.R
\name{HclustGenesHiCor}
\alias{HclustGenesHiCor}
\title{Find the most highly correlated genes using hierarchical clustering}
\usage{
HclustGenesHiCor(gem, seed, cuts)
}
\arguments{
\item{gem}{Gene expression matrix}

\item{seed}{Seed of highly correlating samples}

\item{cuts}{Number of groups to cut dendogram into}
}
\value{
Numeric vector of most highly correlated genes
}
\description{
Find the most highly correlated genes using hierarchical clustering
}

