% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/HclustGenesHiCor.R
\name{HclustGenesHiCor}
\alias{HclustGenesHiCor}
\title{Find the most highly correlated genes using hierarchical clustering}
\usage{
HclustGenesHiCor(gem, seed, cuts)
}
\arguments{
\item{gem}{}

\item{seed}{}

\item{cuts}{}
}
\value{
Numeric vector of most highly correlated genes
}
\description{
Find the most highly correlated genes using hierarchical clustering
}

