% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MannWhitneyGOTerms.R
\name{MannWhitneyGOTerms}
\alias{MannWhitneyGOTerms}
\title{Runs Mann-Whitney test for all GO terms}
\usage{
MannWhitneyGOTerms(genes, gene.values)
}
\arguments{
\item{genes}{}
}
\value{
p-value for each GO Term
}
\description{
Runs Mann-Whitney test for all GO terms
}

