% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiSampleSort.R
\name{MultiSampleSort}
\alias{MultiSampleSort}
\title{SampleSort for multiple patterns}
\usage{
MultiSampleSort(gem, av.corvec, top.genes.num, groups, initial.seeds, num.cores,
  sort.length = NULL)
}
\arguments{
\item{gem}{}
}
\value{
Order of samples by strength to correlation pattern
}
\description{
SampleSort for multiple patterns
}

