% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MultiSampleSort.R
\name{MultiSampleSort}
\alias{MultiSampleSort}
\title{SampleSort for multiple patterns}
\usage{
MultiSampleSort(gem, av.corvec, top.genes.num, groups, initial.seeds, num.cores,
  sort.length)
}
\arguments{
\item{gem}{}

\item{av.corvec}{}

\item{top.genes.num}{}

\item{groups}{}

\item{initial.seeds}{}

\item{num.cores}{}

\item{sort.length}{}
}
\value{
Order of samples by strength to correlation pattern
}
\description{
SampleSort for multiple patterns
}

