% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PC1VecFun.R
\name{PC1VecFun}
\alias{PC1VecFun}
\title{Calculate PC1 of found pattern}
\usage{
PC1VecFun(top.gem, seed.sort, n)
}
\arguments{
\item{top.gem}{Gene expression matrix containing only highly correlating genes}

\item{seed.sort}{Ordering of samples according to strength of correlation}

\item{n}{Number of samples to use in calculation of PC1}
}
\value{
PC1 value for each sample
}
\description{
Calculate PC1 of found pattern
}

