% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointScoreCalc.R
\name{PointScoreCalc}
\alias{PointScoreCalc}
\title{Calculate point score for samples based on gene groups}
\usage{
PointScoreCalc(gem, gloc1, gloc2)
}
\arguments{
\item{gem}{Gene expression matrix}

\item{gloc1}{Location of genes highly correlated to each other but anti-correlated to group2}

\item{gloc2}{Location of genes highly correlated to each other but anti-correlated to group1}
}
\value{
Numeric vector of point score for each sample
}
\description{
Calculate point score for samples based on gene groups
}

