% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SampleSort.R
\name{SampleSort}
\alias{SampleSort}
\title{Sort all samples by strength to correlation pattern}
\usage{
SampleSort(gem, seed, num.cores, sort.length = NULL)
}
\arguments{
\item{gem}{}

\item{seed}{}

\item{num.cores}{}

\item{sort.length}{}
}
\value{
Order of samples by strength to correlation pattern
}
\description{
Sort all samples by strength to correlation pattern
}

