% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleSort.R
\name{SampleSort}
\alias{MultiSampleSortPrep}
\alias{SampleSort}
\title{Methods for ordering samples}
\usage{
SampleSort(gem, seed, num.cores = NULL, sort.length = NULL)

MultiSampleSortPrep(gem, av.corvec, top.genes.num, groups, initial.seeds)
}
\arguments{
\item{gem}{Gene expression matrix}

\item{seed}{Sample seed of highly correlating genes}

\item{num.cores}{Number of cores used in parallel evaluation}

\item{sort.length}{Number of samples to be sorted}

\item{av.corvec}{List of average correlation vector}

\item{top.genes.num}{Number of the top genes in correlation vector to use for sorting samples}

\item{groups}{List showing what runs belong to which correlation vector group}

\item{initial.seeds}{List of sample seeds from all runs}
}
\value{
Order of samples by strength to correlation pattern
}
\description{
Methods for ordering samples
}
\examples{
data(CCLE_small)
data(Mitochondrial_genes)

mito.loc <- (row.names(CCLE_small) \%in\% Mitochondrial_genes)
CCLE.mito <- CCLE_small[mito.loc,]

set.seed(102)
CCLE.seed <- FindSeed(gem = CCLE.mito,
                      seed.size = 10,
                      iterations = 100,
                      messages = 1000)

CCLE.sort <- SampleSort(gem = CCLE.mito,seed = CCLE.seed,sort.length = 11)

# Full ordering can be loaded here:
data("Vignette_sort")
CCLE.samp.sort <- Vignette_sort[[1]]

CCLE.pc1 <- PC1VecFun(top.gem = CCLE.mito,
                      seed.sort = CCLE.samp.sort,
                      n = 10)

CCLE.cor.vec <-  CVEval(gem.part = CCLE.mito,
                            gem.all = CCLE_small,
                            seed = CCLE.seed,
                            splits = 10)

CCLE.bic <- ThresholdBic(cor.vec = CCLE.cor.vec,sort.order = CCLE.samp.sort,
                         pc1 = as.numeric(CCLE.pc1))

CCLE.pc1 <- PC1Align(gem = CCLE_small, pc1 = CCLE.pc1,
                     cor.vec = CCLE.cor.vec ,
                     sort.order = CCLE.samp.sort,
                     bic =CCLE.bic)

CCLE.fork <- ForkClassifier(CCLE.pc1, samp.num = length(CCLE.bic[[2]]))

}

