% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleSortGroups.R
\name{SampleSortGroups}
\alias{SampleSortGroups}
\title{Sort all samples by strength to a known correlation pattern}
\usage{
SampleSortGroups(gem, group1.loc, group2.loc, seed, num.cores = NULL,
  sort.length = NULL)
}
\arguments{
\item{gem}{}
}
\value{
Order of samples by strength to correlation pattern
}
\description{
Sort all samples by strength to a known correlation pattern
}

