% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeedGenerator.R
\name{SeedGenerator}
\alias{SeedGenerator}
\title{Calculate initial seeds to minimise overlaps}
\usage{
SeedGenerator(seed.size, numbers, sample.length, break.num = 0,
  attempts = 1000)
}
\value{
Initial seed list
}
\description{
Calculate initial seeds to minimise overlaps
}

