% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SilhouetteClustGroups.R
\name{SilhouetteClustGroups}
\alias{SilhouetteClustGroups}
\title{Slihouette validation of correlation vector clusters}
\usage{
SilhouetteClustGroups(cor.vec.mat, max.clusters, plots = FALSE, seed1 = 100,
  rand.vec = TRUE)
}
\value{
The distinct clusters of correlation vectors
}
\description{
Slihouette validation of correlation vector clusters
}

