% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThresholdBic.R
\name{ThresholdBic}
\alias{ThresholdBic}
\title{Calculate bicluster threshold}
\usage{
ThresholdBic(cor.vec, sort.order, pc1, samp.sig = 0)
}
\arguments{
\item{cor.vec}{Correlation vector}

\item{sort.order}{Order of samples}

\item{pc1}{PC1 values for samples}
}
\value{
Genes and samples in bicluster
}
\description{
Calculate bicluster threshold
}

