% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThresholdBic.R
\name{ThresholdBic}
\alias{ThresholdBic}
\title{Calculate bicluster threshold}
\usage{
ThresholdBic(cor.vec, sort.order, pc1, samp.sig = 0)
}
\value{
Genes and samples in bicluster
}
\description{
Calculate bicluster threshold
}

